refactor(core-weave): extract payload version layout planning
- move payload history/state/manifestation layout resolution into payload_version_layout.ts
- move guarded overwriteExistingState validation and planning into payload_overwrite.ts
- move shared artifact manifestation path construction into artifact_manifestation_paths.ts
- keep payload renderers, ResourcePage builders, shape assertions, and planner façade behavior in weave.ts
- verify with fmt, lint, check, core weave tests, integration weave tests, and import-cycle audit
